requirements.txt
# 数据获取和处理
akshare==1.16.74
numpy==1.26.2
pandas==2.2.3
tushare==1.4.13

# 数据库相关
pymysql==1.1.1
sqlalchemy==2.0.35

# 工具包
python-dateutil==2.9.0.post0
requests==2.32.3
schedule==1.2.2
tqdm==4.66.5
xlrd==2.0.1

# Redis相关
redis>=5.0.0

# 异步HTTP
aiohttp==3.11.13
aiodns==3.1.1
async-timeout==4.0.3


openai==1.12.0

# 工具库
python-dotenv>=1.0.0

markdown==3.8
